Original version song
We’ve added the original version of the song “Австралия” by the 90s Russian pop group МАЙАМИ to TTBus Radio!

This track, released in the late 1990s, is all about the cold Russian winter and the dream of escaping to somewhere warm and sunny — like Australia. It reflects the feeling of wanting to leave behind icy streets and frozen eyelashes for kangaroos, beaches, and sunshine.

We discovered this song recently and it fits perfectly with the vibes of a chill escape. That’s why we’re using it as the original version for Kékhegy's TTBus Radio lineup.